# lumina-api/swagger.yaml
openapi: 3.0.3
info:
  title: Lumina Bot API
  version: 3.0.0
  contact:
    name: Lumina Dev Team
    url: https://bot.luminasink.com
  description: |
    # Lumina Bot API

    REST API for the Lumina Bot dashboard, inventory/gacha system, Discord guild
    management, and the internal bot ↔ API bridge. Built on Node.js/Express,
    deployed to Vercel (production) and `localhost:3000` (dev).

    ---

    ## Authentication

    The API supports **two** authentication methods for user-facing routes:

    1. **Cookie `lumina_token` (preferred for browsers)**
       - `httpOnly`, `Secure` (prod), `SameSite=Lax`, max-age = 1 hour
       - Set by `POST /expapi/v1/login` and the OAuth2 callback
       - Sent automatically by the browser — **immune to XSS** (JavaScript cannot read it)

    2. **Header `Authorization: Bearer <jwt>` (legacy / non-browser)**
       - For CLI tools, scripts, integration tests, and any client that does not
         manage cookies
       - Same JWT, same 1h expiry

    JWT expiry: **1 hour**. There is no refresh-token rotation yet — clients must
    re-authenticate when the token expires.

    The session helper `verifyRequestAuth(req)` reads the cookie first, then
    falls back to the Bearer header. Either is accepted on every protected route.

    ---

    ## CSRF Protection

    State-changing routes (`POST`/`PUT`/`DELETE` with `csrfProtectionNeeded: true`)
    require the **double-submit cookie** pattern:

...

...
    rotated hourly.

...

    ---

...

...

...

    ---
...
servers:
...
components:
  securitySchemes:

    cookieAuth:
      type: apiKey
      in: cookie
      name: lumina_token
      description: |
        httpOnly JWT cookie set by `POST /expapi/v1/login` or the OAuth2 callback.
        Preferred auth method for browsers (immune to XSS). Sent automatically by
        the browser; non-browser clients should use `bearerAuth` instead.

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <jwt>` — legacy/alternative auth for non-browser
        clients (CLI, scripts, integration tests). Same JWT, same 1h expiry as
        the cookie.

    internalKey:
      type: apiKey
      in: header
      name: internal-key
      description: |
        Required on `/expapi/internal/*` routes. Constant-time compared against
        `INTERNAL_API_KEY`. Only the Discord bot should possess this key.

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on state-changing routes (`POST`/`PUT`/`DELETE`) that have
        `csrfProtectionNeeded: true`. The value must match the `csrf_token`
        cookie obtained from `GET /expapi/v1/csrf-token`.

    apiKeyQuery:
      type: apiKey
      in: query
      name: apiKey
...
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message (PT-BR).
        code:
          type: string
          description: Stable machine-readable error code.
          example: INVALID_CREDENTIALS
        details:
          type: object
          additionalProperties: true
          description: Optional structured details (e.g. attemptsRemaining, lockedUntil).
...
    DiscordGuild:
      type: object
      description: Discord guild shape returned by `/discord/guild/:guildId`.
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
        member_count:
          type: integer
          nullable: true
        approximate_member_count:
          type: integer
          nullable: true
        approximate_presence_count:
          type: integer
        owner_id:
          type: string
        features:
          type: array
          items:
            type: string
        verification_level:
          type: integer
        premium_tier:
          type: integer
        roles:
          type: array
          items:
            type: object
        system_channel_id:
          type: string
          nullable: true
        nsfw:
          type: boolean
        nsfw_level:
          type: integer
        preferred_locale:
          type: string
        mfa_level:
          type: integer
        max_members:
          type: integer
          nullable: true
        emojis:
          type: array
          items:
            type: object
        stickers:
          type: array
          items:
            type: object
        premium_progress_bar_enabled:
          type: boolean
        splash:
          type: string
          nullable: true
        banner:
          type: string
          nullable: true
...
  responses:
    Unauthorized:
      description: Authentication missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Token inválido ou expirado.
            code: INVALID_TOKEN
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Conta não encontrada.
            code: ACCOUNT_NOT_FOUND
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Erro interno do servidor.
            code: UNHANDLED_ERROR
paths:
...
  /expapi/v1/discordinfo:
    get:
      tags: [Discord]
      summary: Get Discord user info
      description: |
        Returns the linked Discord account's public info (`avatar`, `username`,
        `id`) by calling Discord's `/users/@me` with the stored OAuth2 token.

        Refreshes the Discord OAuth2 token automatically if it has expired.
      operationId: getDiscordInfo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Discord info retrieved.
          content:
            application/json:
              schema:
                type: object
                properties:
                  avatar:
                    type: string
                    description: Discord avatar hash (use CDN to build URL).
                  username:
                    type: string
                  id:
                    type: string
                    description: Discord user ID.
        '400':
          description: Discord not linked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Conta Discord não vinculada.
                code: DISCORD_NOT_LINKED
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /expapi/v1/discord/guild/{guildId}:
    get:
      tags: [Discord]
      summary: Get guild info from Discord
      description: |
        Returns full Discord guild info for the given `guildId`, fetched via
        the bot token (with `with_counts=true`).

        **Access control:**
        - User must have Discord linked
        - User must be a member of the guild (verified via `/users/@me/guilds`)
          — prevents IDOR

        Refreshes the user's Discord OAuth2 token automatically if expired.

        Accepts either the `lumina_token` cookie or the Bearer header.
      operationId: getDiscordGuildInfo
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
            pattern: '^\d{17,19}$'
      responses:
        '200':
          description: Guild info retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordGuild'
        '400':
          description: Invalid guild ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Guild ID inválido.
                code: INVALID_GUILD_ID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            Discord not linked, token expired/refresh failed, missing `guilds`
            scope, or user is not a member of this guild.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_linked:
                  value:
                    error: Você precisa vincular sua conta Discord para acessar esta funcionalidade.
                    code: DISCORD_NOT_LINKED
                not_member:
                  value:
                    error: Você não é membro desta guilda.
                    code: NOT_GUILD_MEMBER
                token_expired:
                  value:
                    error: Sua conexão com o Discord expirou. Reconecte sua conta nas configurações.
                    code: DISCORD_TOKEN_EXPIRED
                missing_scope:
                  value:
                    error: Seu token Discord não tem permissão para listar servidores. Reconecte sua conta.
                    code: DISCORD_MISSING_GUILDS_SCOPE
        '404':
          description: Guild not found (or bot is not in it).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Guilda não encontrada ou o bot não está neste servidor.
                code: GUILD_NOT_FOUND
        '500':
          $ref: '#/components/responses/ServerError'
        '502':
          description: Discord API error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Erro ao verificar seus servidores no Discord.
                code: DISCORD_API_ERROR
        '503':
          description: Bot token missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing:
                  value:
                    error: O bot Discord ainda não foi configurado no painel administrativo.
                    code: BOT_TOKEN_MISSING
                invalid:
                  value:
                    error: O token do bot Discord configurado é inválido.
                    code: BOT_TOKEN_INVALID
        '504':
          description: Discord timeout.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Tempo esgotado ao contatar o Discord. Tente novamente.
                code: DISCORD_TIMEOUT

  # ───────────────────────────────────────────────────────────────────────
  # ADMIN
  # ───────────────────────────────────────────────────────────────────────
...